fix add middleware
